feat: Implement btreeAdapter for CRUD operations in litegodb
